--- data/league_standings_away.csv
@@ -1,21 +1,0 @@
-Position,Team,Matches,Wins,Draws,Losses,Goals For,Goals Against,Points,Promotion/Relegation
-1,Real Madrid,18,13,4,1,35,13,43,Champions League
-2,Barcelona,18,10,6,2,34,22,36,Champions League
-3,Girona FC,18,9,4,5,29,25,31,Champions League
-4,Real Sociedad,18,7,6,5,23,19,27,UEFA Europa League
-5,Villarreal,18,7,5,6,28,32,26,N/A
-6,Atlético Madrid,18,7,3,8,26,21,24,Champions League
-7,Athletic Club,18,6,5,7,18,19,23,UEFA Europa League
-8,Real Betis,18,5,7,6,21,26,22,UEFA Conference League Qualification
-9,Osasuna,18,5,4,9,22,29,19,N/A
-10,Sevilla,18,4,6,8,21,25,18,N/A
-11,Rayo Vallecano,18,4,6,8,11,19,18,N/A
-12,Valencia,18,5,3,10,18,29,18,N/A
-13,Las Palmas,18,4,3,11,13,27,15,N/A
-14,Getafe,18,2,8,8,22,31,14,N/A
-15,Celta Vigo,18,3,5,10,23,33,14,N/A
-16,Deportivo Alavés,18,3,5,10,12,26,14,N/A
-17,Mallorca,18,1,8,9,14,27,11,N/A
-18,Almería,18,2,3,13,19,37,9,Relegation
-19,Cádiz,18,1,6,11,9,30,9,Relegation
-20,Granada,18,0,3,15,14,40,3,Relegation
