--- data/league_standings_home.csv
@@ -1,21 +1,0 @@
-Position,Team,Matches,Wins,Draws,Losses,Goals For,Goals Against,Points,Promotion/Relegation
-1,Real Madrid,18,16,2,0,48,9,50,Champions League
-2,Atlético Madrid,18,16,1,1,41,18,49,Champions League
-3,Girona FC,18,14,2,2,46,20,44,Champions League
-4,Barcelona,18,14,1,3,40,21,43,Champions League
-5,Athletic Club,18,11,6,1,40,18,39,UEFA Europa League
-6,Real Betis,18,9,7,2,27,17,34,UEFA Conference League Qualification
-7,Real Sociedad,18,8,6,4,26,18,30,UEFA Europa League
-8,Valencia,18,8,6,4,19,11,30,N/A
-9,Getafe,18,8,5,5,19,20,29,N/A
-10,Deportivo Alavés,18,8,4,6,22,19,28,N/A
-11,Villarreal,18,7,4,7,32,28,25,N/A
-12,Mallorca,18,6,7,5,15,14,25,N/A
-13,Sevilla,18,6,5,7,26,25,23,N/A
-14,Las Palmas,18,6,5,7,19,19,23,N/A
-15,Celta Vigo,18,6,5,7,19,21,23,N/A
-16,Cádiz,18,5,8,5,16,19,23,Relegation
-17,Osasuna,18,6,4,8,18,25,22,N/A
-18,Rayo Vallecano,18,4,8,6,18,25,20,N/A
-19,Granada,18,4,6,8,23,30,18,Relegation
-20,Almería,18,0,8,10,16,35,8,Relegation
