--- data/league_standings_total.csv
@@ -1,21 +1,0 @@
-Position,Team,Matches,Wins,Draws,Losses,Goals For,Goals Against,Points,Promotion/Relegation
-1,Real Madrid,36,29,6,1,83,22,93,Champions League
-2,Barcelona,36,24,7,5,74,43,79,Champions League
-3,Girona FC,36,23,6,7,75,45,75,Champions League
-4,Atlético Madrid,36,23,4,9,67,39,73,Champions League
-5,Athletic Club,36,17,11,8,58,37,62,UEFA Europa League
-6,Real Sociedad,36,15,12,9,49,37,57,UEFA Europa League
-7,Real Betis,36,14,14,8,48,43,56,UEFA Conference League Qualification
-8,Villarreal,36,14,9,13,60,60,51,N/A
-9,Valencia,36,13,9,14,37,40,48,N/A
-10,Getafe,36,10,13,13,41,51,43,N/A
-11,Deportivo Alavés,36,11,9,16,34,45,42,N/A
-12,Sevilla,36,10,11,15,47,50,41,N/A
-13,Osasuna,36,11,8,17,40,54,41,N/A
-14,Las Palmas,36,10,8,18,32,46,38,N/A
-15,Rayo Vallecano,36,8,14,14,29,44,38,N/A
-16,Celta Vigo,36,9,10,17,42,54,37,N/A
-17,Mallorca,36,7,15,14,29,41,36,N/A
-18,Cádiz,36,6,14,16,25,49,32,Relegation
-19,Granada,36,4,9,23,37,70,21,Relegation
-20,Almería,36,2,11,23,35,72,17,Relegation
