--- data/team_statistics.csv
@@ -1,21 +1,0 @@
-Team,Goals Scored,Goals Conceded,Assists,Shots,Penalty Goals,Penalties Taken,Free Kick Goals,Goals From Inside The Box,Goals From Outside The Box,Yellow Cards,Red Cards
-Barcelona,74,43,51,570,6,7,2,65,6,81,2
-Real Madrid,83,22,62,573,2,5,0,72,9,64,4
-Deportivo Alavés,34,45,26,442,3,6,0,26,5,78,1
-Las Palmas,32,46,22,373,3,5,1,24,8,78,5
-Granada,37,70,27,381,5,7,0,31,3,91,3
-Girona FC,75,45,57,450,5,6,0,69,6,72,1
-Atlético Madrid,67,39,46,457,5,6,0,59,6,79,4
-Athletic Club,58,37,38,441,3,5,0,49,7,76,5
-Real Sociedad,49,37,35,442,4,4,1,41,5,91,2
-Real Betis,48,43,31,463,2,2,1,38,8,82,3
-Villarreal,60,60,43,437,5,7,0,58,1,101,8
-Valencia,37,40,24,345,6,7,0,34,3,60,5
-Getafe,41,51,26,427,6,7,0,40,1,125,8
-Sevilla,47,50,29,465,2,2,0,38,7,105,4
-Osasuna,40,54,27,399,4,5,0,35,5,71,5
-Mallorca,29,41,23,380,1,3,0,27,2,99,3
-Celta Vigo,42,54,32,455,2,4,0,35,6,63,5
-Rayo Vallecano,29,44,12,461,6,6,1,24,5,101,5
-Cádiz,25,49,16,376,3,3,1,21,3,101,6
-Almería,35,72,22,446,2,2,0,32,3,84,6
