+++ data_extraction/__init__.py
@@ -1,0 +1,5 @@
+# data_extraction/__init__.py
+
+from .CsvDataExtractor import CsvDataExtractor
+from .DataExtractor import DataExtractor
+from .SofascoreDataExtractor import SofascoreDataExtractor
