+++ run_tests.py
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+python -m unittest discover tests
